add collabsible menu/tray component
